Solved Problem 1: https://projecteuler.net/problem=1
Answer: 233168

I could have used a simple loop of n from 1 to 1000 and if n % 3|5 = 0 then add
the number but that would be less efficient than tracing the pattern of what is
added each time and going from there. While with n only going up to 1000 this
is trivial, I wanted as optimized a solution as I could possibly get.